+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.7 - Unreleased
 
   * Fix `Collectable` implementation silently ignoring write errors.
+  * Fix `Collectable` implementation silently ignoring close errors.
 
 ## v0.2.6 - 2026-05-01